supervisor process tweak so process is stoppable
Don't know exacly whats going on but while enhancing your chat example so it would actually halt (via SIGTERM signal, from supervisor) I noticed that the php process was't killed at all (resulting in a port in use error on the next start). The 'SIGTERM' signal never reached PHP.

After googling and fiddling a bit this seems to be a better way. the signal arrives in PHP and the process is halted. and the ulimit seems to be in effect (tested by using 1)